+++ Dashboard/extract_from_patagonia.py
@@ -1,23 +1,25 @@
-"Extract Script: Pulls product information from ASOS webspage"
+"Extract Script: Pulls product information from Patagonia webpage"
 import json
 import logging
 from bs4 import BeautifulSoup
-from utils.helpers import get_soup, configure_logging
+from helpers import get_soup, configure_logging
 
 
 def is_correct_page(soup: BeautifulSoup) -> bool:
-    """Returns True if asos product page, else False."""
+    """Returns True if patagonia product page, else False."""
     if not isinstance(soup, BeautifulSoup):
+        logging.error("Soup must be of type BeautifulSoup")
         raise TypeError('Soup must be of type BeautifulSoup')
 
     single_product_identifier = soup.find(
-        'div', attrs={'class': 'single-product'})
+        'div', attrs={'class': 'product-detail'})
     return single_product_identifier is not None
 
 
 def scrape_product_information(soup: BeautifulSoup) -> dict | None:
     """Extract product information from a BeautifulSoup object."""
     if not isinstance(soup, BeautifulSoup):
+        logging.error("Soup must be of type BeautifulSoup")
         raise TypeError('Soup must be of type BeautifulSoup')
 
     product_soup = soup.find('script', type="application/ld+json")
@@ -28,7 +30,7 @@ def scrape_product_information(soup: BeautifulSoup) -> dict | None:
     return None
 
 
-def get_product_code_asos(product_data: dict) -> str | None:
+def get_product_code_patagonia(product_data: dict) -> str | None:
     """Returns product ID from the webpage"""
     if not isinstance(product_data, dict):
         raise TypeError('product_info must be of type dict')
@@ -36,21 +38,20 @@ def get_product_code_asos(product_data: dict) -> str | None:
         logging.error("Missing product data")
         return None
 
-    product_id = product_data.get("productID")
+    product_id = product_data.get("mpn")
     if product_id:
         return str(product_id)
 
-    graph = product_data.get("@graph")
-    if graph:
-        product_id = graph[0]['productID']
-        return product_id
+    sku = product_data.get("sku")
+    if sku:
+        return str(sku)
 
     logging.error("Missing productID in product_data")
     return None
 
 
-def get_product_name_asos(product_data: dict) -> str | None:
-    """Returns product ID from the webpage"""
+def get_product_name_patagonia(product_data: dict) -> str | None:
+    """Returns product name from the webpage"""
     if not isinstance(product_data, dict):
         raise TypeError('product_info must be of type dict')
 
@@ -62,12 +63,7 @@ def get_product_name_asos(product_data: dict) -> str | None:
     if product_name:
         return product_name
 
-    graph = product_data.get("@graph")
-    if graph:
-        product_name = graph[0]['name']
-        return product_name
-
-    logging.error("Missing productID in product_data")
+    logging.error("Missing name in product_data")
     return None
 
 
@@ -94,13 +90,14 @@ def extract_product_information(url: str) -> dict | None:
         logging.error("Unable to extract information from product page!")
         return None
 
-    product_code = get_product_code_asos(data)
-    product_name = get_product_name_asos(data)
+    product_code = get_product_code_patagonia(data)
+    product_name = get_product_name_patagonia(data)
     if not (product_code and product_name):
         logging.error(
             'Unable to get correct product code or product name from website!')
         return None
     logging.info("Extraction completed successfully!")
+
     return {
         'url': url,
         'product_code': product_code,
