+++ Dashboard/login.py
@@ -11,7 +11,6 @@ from psycopg2 import connect
 from psycopg2.extensions import connection
 
 from navigation import make_sidebar
-from utils.email_verification import get_ses_client, send_verification_email
 
 EMAIL_PATTERN = r"""(?:[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*|\"(?:[\x01-\x08\x0b\x0c\x0e-\x1f\x21\x23-\x5b\x5d-\x7f]|\\[\x01-\x09\x0b\x0c\x0e-\x7f])*\")@(?:(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?|\[(?:(?:(2(5[0-5]|[0-4][0-9])|1[0-9][0-9]|[1-9]?[0-9]))\.){3}(?:(2(5[0-5]|[0-4][0-9])|1[0-9][0-9]|[1-9]?[0-9])|[a-z0-9-]*[a-z0-9]:(?:[\x01-\x08\x0b\x0c\x0e-\x1f\x21-\x5a\x53-\x7f]|\\[\x01-\x09\x0b\x0c\x0e-\x7f])+)\])"""
 
@@ -85,8 +84,6 @@ def login_page(config: _Environ, email_pattern: str) -> None:
             st.error(
                 f"The email {signup_email} is already registered. Please log in.")
         elif re.match(email_pattern, signup_email):
-            client = get_ses_client(config)
-            send_verification_email(client, signup_email)
             add_email(conn, signup_email)
             login(signup_email)
         else:
