+++ Dashboard/pages/price.py
@@ -6,15 +6,14 @@ from datetime import datetime
 
 from dotenv import load_dotenv
 import streamlit as st
+import streamlit_nested_layout
 import pandas as pd
 import altair as alt
 from psycopg2.extensions import connection
 
-from navigation import make_sidebar
 
-from utils.helpers import get_cursor, get_connection
-from utils.load import load_product_data
-from utils.ses_get_emails import get_ses_client, get_ses_emails
+from navigation import make_sidebar
+from helpers import get_cursor, get_connection
 
 
 def can_parse_as_float(to_check: any) -> bool:
@@ -28,12 +27,6 @@ def can_parse_as_float(to_check: any) -> bool:
         return False
 
 
-def get_supported_websites(conn: connection) -> list:
-    """Gets a list of supported websites."""
-    with get_cursor(conn) as cur:
-        cur.execute("SELECT website_name from websites;")
-        websites = cur.fetchall()
-    return [website["website_name"] for website in websites]
 
 
 def get_user_id(conn: connection, email: str) -> int | None:
@@ -43,86 +36,6 @@ def get_user_id(conn: connection, email: str) -> int | None:
         user_id = cur.fetchone()
     return user_id.get("user_id")
 
-
-def get_product_id(conn: connection, url: str) -> int | None:
-    """Gets the products id from the database given a url."""
-    with get_cursor(conn) as cur:
-        cur.execute("SELECT product_id from products WHERE url = %s;", (url,))
-        product_id = cur.fetchone()
-    return product_id.get("product_id")
-
-
-def is_subscription_in_table(conn: connection, user_id: int, product_id: int) -> bool:
-    """Determines if a user is subscribed to a particular product."""
-    with get_cursor(conn) as cur:
-        cur.execute("""SELECT * FROM subscriptions
-                        WHERE user_id = %s AND product_id = %s""",
-                    (user_id, product_id))
-        counts = cur.fetchone()
-    return counts is not None
-
-
-def insert_subscription_into_db(
-        conn: connection, user_id: int, product_id: int,
-        price_threshold: float | int | None) -> bool:
-    """Inserts a new subscription into the subscriptions table.
-    Returns True on success, else False."""
-    try:
-        with get_cursor(conn) as cur:
-            cur.execute("""INSERT INTO subscriptions
-                         (user_id, product_id, price_threshold)
-                        VALUES (%s, %s, %s);""", (user_id, product_id, price_threshold))
-            conn.commit()
-        return True
-    except Exception as e:
-        st.write(e)
-        return False
-
-
-def update_db_to_subscribe(
-        config: _Environ, product_url: str, price_threshold: bool | None, user_email: str) -> bool:
-    """Inserts a subscription into a database, if it doesn't already exist.
-    Returns True on success, else False"""
-    conn = get_connection(config)
-    user_id = get_user_id(conn, user_email)
-    if not user_id:
-        logging.error("User not validated. Please try again!")
-        st.warning("User not validated. Please try again!")
-        return False
-    product_id = get_product_id(conn, product_url)
-    if not product_id:
-        logging.error("Unexpected error, please try again later.")
-        st.warning("Unexpected error, please try again later.")
-        return False
-    if is_subscription_in_table(conn, user_id, product_id):
-        logging.error("You are already subscribed to this product!")
-        st.warning("You are already subscribed to this product!")
-        return True
-    logging.info("New subscription added!")
-    st.success("New subscription added!")
-    return insert_subscription_into_db(conn, user_id, product_id, price_threshold)
-
-
-def subscribe_to_product(config: _Environ, product_url: str, price_threshold: bool | None) -> bool:
-    """adds a product to the database and return True if successful
-    (or if product is already in the database), else False."""
-    try:
-        load_product_data(config, product_url)
-    except (TypeError, ValueError):
-        logging.error(
-            "Invalid URL, please enter a URL from one of the supported websites!")
-        st.warning(
-            "Invalid URL, please enter a URL from one of the supported websites!")
-        return False
-    try:
-        if price_threshold == "":
-            price_threshold = None
-        return update_db_to_subscribe(config, product_url,
-                                      price_threshold, st.session_state["email"])
-    except (TypeError, ValueError):
-        logging.error("Unable to subscribe to product!")
-        st.warning("Unable to subscribe to product!")
-        return False
 
 
 def get_subscribed_products(config: _Environ, email: str) -> list[dict]:
@@ -191,7 +104,7 @@ def change_threshold_in_db(config: _Environ, new_threshold: float | None, produc
         conn.commit()
     
 
-def change_threshold(config: _Environ, product_information: dict, column) -> None:
+def change_threshold(config: _Environ, product_information: dict) -> None:
     text_input_placeholder = None
     if product_information['price_threshold'] is not None:
         text_input_placeholder = f"£{float(product_information['price_threshold']):.2f}"
@@ -199,25 +112,25 @@ def change_threshold(config: _Environ, product_information: dict, column) -> Non
                                     placeholder=text_input_placeholder,
                                     help='Can be left empty to remove threshold!',
                                     key=f'new_threshold_value_{product_information['product_id']}')
-    with column:
-        if st.button('Update Threshold', key=f'new_threshold_submit_{product_information['product_id']}'):
-            valid_threshold = False
-            if new_threshold == '':
-                new_threshold = None
+    st.write('')
+    if st.button('Update Price Alert', key=f'new_threshold_submit_{product_information['product_id']}'):
+        valid_threshold = False
+        if new_threshold == '':
+            new_threshold = None
+            valid_threshold = True
+        elif can_parse_as_float(new_threshold):
+            new_threshold = float(new_threshold)
+            if new_threshold > 0:
                 valid_threshold = True
-            elif can_parse_as_float(new_threshold):
-                new_threshold = float(new_threshold)
-                if new_threshold > 0:
-                    valid_threshold = True
-            if not valid_threshold and new_threshold == product_information['price_threshold']:
-                st.warning('New threshold is the same as the old threshold.')
-            elif valid_threshold:
-                change_threshold_in_db(config, new_threshold, product_information['product_id'], st.session_state['email'])
-                st.rerun()
-            elif not valid_threshold and isinstance (new_threshold, float):
-                st.warning('Threshold must be a positive number!')
-            else:
-                st.warning('Invalid threshold, please try again!')
+        if not valid_threshold and new_threshold == product_information['price_threshold']:
+            st.warning('New threshold is the same as the old threshold.')
+        elif valid_threshold:
+            change_threshold_in_db(config, new_threshold, product_information['product_id'], st.session_state['email'])
+            st.rerun()
+        elif not valid_threshold and isinstance (new_threshold, float):
+            st.warning('Threshold must be a positive number!')
+        else:
+            st.warning('Invalid threshold, please try again!')
 
 def unsubscribe_from_product_in_db(config: _Environ, product_id: int, email: str) -> None:
     conn = get_connection(config)
@@ -228,27 +141,49 @@ def unsubscribe_from_product_in_db(config: _Environ, product_id: int, email: str
                         AND user_id = %s''', (product_id, user_id))
         conn.commit()
 
-def unsubscribe_from_product(config: _Environ, product_information: int, column) -> None:
-    with column:
-        if st.button('Unsubscribe', key=f'unsubscribe_button_{product_information['product_id']}'):
-            unsubscribe_from_product_in_db(config, product_information['product_id'], st.session_state['email'])
-            st.rerun()
+def unsubscribe_from_product(config: _Environ, product_information: int) -> None:
+    if st.button('Unsubscribe', key=f'unsubscribe_button_{product_information['product_id']}'):
+        unsubscribe_from_product_in_db(config, product_information['product_id'], st.session_state['email'])
+        st.rerun()
+
+def product_price_change(price_readings: pd.DataFrame) -> int:
+    '''returns an emoji representing the emoji regarding the price.'''
+    if len(price_readings) == 1:
+        return 0
+    price_readings_sorted_by_time = price_readings.sort_values(by='reading_at', ascending=False)[:2].reset_index()
+    if price_readings_sorted_by_time.loc[0]['price'] > price_readings_sorted_by_time.loc[1]['price']:
+        return 1
+    elif price_readings_sorted_by_time.loc[0]['price'] == price_readings_sorted_by_time.loc[1]['price']:
+        return 0
+    return -1
+
+def get_price_change_emoji(price_change_enum: int) -> str:
+    if price_change_enum == 0:
+        return '➖'
+    elif price_change_enum == 1:
+        return '📈'
+    return '📉'
+
+def get_price_change_colour(price_change_enum: int) -> str:
+    if price_change_enum == 0:
+        return 'orange'
+    elif price_change_enum == 1:
+        return 'red'
+    return 'green'
 
 def display_subscribed_product(config: _Environ, product_information: dict) -> None:
     """For a particular product, get the price readings and display a container
     for that product, containing relevant information."""
     price_readings = get_price_readings(
         config, product_information["product_id"])
-
-    with st.container(border=True):
+    expander = st.expander(f'Fetching data for {product_information['product_name']}')
+    if price_readings is not None:
+        current_price = float(price_readings[
+                    price_readings['reading_at'] == price_readings['reading_at'].max()]['price'])
+        price_change_enum = product_price_change(price_readings)
+        expander = st.expander(f'**£:{get_price_change_colour(price_change_enum)}[{current_price:.2f}]** - {product_information['product_name']}', icon=get_price_change_emoji(price_change_enum))
+    with expander:
         st.write('')
-        header_col1, header_col2 = st.columns([1, 19])
-        with header_col1:
-            st.image('logo/hermes_crop.png', width=40)
-        with header_col2:
-            st.write("**%s** - **[%s](%s)**" % (product_information["website_name"].title(),
-                                            product_information["product_name"].title(),
-                                            product_information["url"]))
         if price_readings is not None:
             col1, col2 = st.columns(2)
             
@@ -257,18 +192,20 @@ def display_subscribed_product(config: _Environ, product_information: dict) -> N
             encoded_price_readings = get_encode_price_reading(
                 price_readings, product_information["price_threshold"])
             with col1:
-                
-                current_price = price_readings[
-                    price_readings['reading_at'] == price_readings['reading_at'].max()]['price']
+                inner_col_1, inner_col_2 = st.columns([5,10])
+                with inner_col_1:
+                    st.link_button('Go to Product', url=product_information['url'])
+                with inner_col_2:
+                    unsubscribe_from_product(config, product_information)
                 st.write('')
-                st.write(f'Current Price: £{float(current_price):.2f}')
+                st.write('')
+                st.write('')
+                st.write('')
+                st.write('')
                 st.write('')
                 with st.container(border=True):
                     st.write('')
-                    inner_col1, inner_col2 = st.columns(2)
-                    st.write('')
-                    change_threshold(config, product_information, inner_col1)
-                    unsubscribe_from_product(config, product_information, inner_col2)
+                    change_threshold(config, product_information)
                     st.write('')
             with col2:
                 st.altair_chart(encoded_price_readings)
@@ -278,37 +215,21 @@ def display_subscribed_product(config: _Environ, product_information: dict) -> N
 
 def price_tracker_page(config: _Environ) -> None:
     """Product Price Tracker Page"""
-    conn = get_connection(config)
-    st.title("Product Price Tracker")
-    websites = get_supported_websites(conn)
 
-    with st.expander("List of Supported websites", expanded=True):
-        for website in websites:
-            st.write(website.title())
 
-    with st.expander("Subscribe to new product", expanded=True):
-        with st.form("subscribe_to_product", clear_on_submit=True, border=False):
-            product_url = st.text_input("Enter the product URL:")
-            price_threshold = st.text_input(
-                "Enter a threshold (£):", placeholder="Can be left empty")
-            if st.form_submit_button("Track Product", type="primary"):
-                if not product_url:
-                    logging.error("You must enter a URL.")
-                    st.error("You must enter a URL.")
-                elif price_threshold != "" and not can_parse_as_float(price_threshold):
-                    logging.error("Threshold must be a number (or empty).")
-                    st.warning("Threshold must be a number (or empty).")
-                elif can_parse_as_float(price_threshold) and float(price_threshold) <= 0:
-                    st.warning('Price Threshold must be positive!')
-                else:
-                    subscribe_to_product(config, product_url, price_threshold)
     subscribed_to_products = get_subscribed_products(
         config, st.session_state["email"])
     if subscribed_to_products:
-        st.header('Tracked products:')
+        st.title('Tracked products:')
         subscribed_to_products.sort(key=lambda x:x['product_id'])
-        for product in subscribed_to_products:
-            display_subscribed_product(config, product)
+        websites = list(set(product['website_name'] for product in subscribed_to_products))
+        websites.sort()
+        for website in websites:
+            st.write('')
+            with st.expander(f'**{website}**'.title()):
+                for product in subscribed_to_products:
+                    if product['website_name'] == website:
+                        display_subscribed_product(config, product)
     else:
         st.warning('You are not subscribed to any products, please subscribe above!')
