+++ Dashboard/seeding_data.sql
@@ -1,0 +1,6 @@
+INSERT INTO price_readings (product_id, reading_at, price)
+VALUES
+(7, '2024-06-16 12:18:05', 79.99),
+(7, '2024-06-19 00:29:05', 84.99),
+(7, '2024-06-21 18:09:01', 80.00),
+(7, '2024-06-23 07:27:17', 85.00);
